refactor: ProductList 컴포넌트 삭제 및 무한스크롤 로직 위치 이동
