--- src/components/ProductList/ProductList.styled.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const ProductListContainer = styled.div`
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 2rem 1.6rem;
-`;
-
-export const ObserverContainer = styled.div`
-  width: 100%;
-  height: 1rem;
-`;
