+++ src/pages/Product.styled.ts
@@ -13,3 +13,13 @@ export const ProductTitle = styled.p`
   margin-bottom: 2.4rem;
 `;
 
+export const ProductListContainer = styled.div`
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 2rem 1.6rem;
+`;
+
+export const ObserverContainer = styled.div`
+  width: 100%;
+  height: 1rem;
+`;
