+++ src/pages/Product.tsx
@@ -1,22 +1,39 @@
-import { useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
 import useFetchAddCart from '../hooks/useFetchAddCart';
 import useFetchProducts from '../hooks/useFetchProducts';
-import { CartContext } from '../CartContext';
+import useIntersectionObserver from '../hooks/useIntersectionObserver';
 import Header from '../components/Header/Header';
 import Dropdown from '../components/Dropdown/Dropdown';
 import ErrorMessage from '../components/ErrorMessage/ErrorMessage';
-import ProductList from '../components/ProductList/ProductList';
+import ProductCard from '../components/ProductCard/ProductCard';
+import { CartContext } from '../CartContext';
 import { SortingParam } from '../types/sort';
 
 import * as S from './Product.styled';
 
 function Product() {
+  const target = useRef(null);
+
   const fetchAddCartState = useFetchAddCart();
   const [sortings, setSortings] = useState<SortingParam[]>([]);
   const [filter, setFilter] = useState('');
   const { products, isError, isPending, isLast, fetchNextPage, page } =
     useFetchProducts(sortings, filter);
+
+  const { observe, unobserve } = useIntersectionObserver(() => fetchNextPage());
+
+  useEffect(() => {
+    if (!target.current || isPending || isLast) return;
+    const currentTarget = target.current;
+    observe(currentTarget);
+
+    return () => {
+      if (currentTarget) {
+        unobserve(currentTarget);
+      }
+    };
+  }, [page, isPending, isLast]);
 
   return (
     <>
@@ -26,13 +43,12 @@ function Product() {
         <S.ProductContentWrapper>
           <S.ProductTitle>bpple 상품 목록</S.ProductTitle>
           <Dropdown setSortings={setSortings} setFilter={setFilter} />
-          <ProductList
-            page={page}
-            products={products}
-            fetchNextPage={fetchNextPage}
-            isPending={isPending}
-            isLast={isLast}
-          />
+          <S.ProductListContainer>
+            {products.map((product) => {
+              return <ProductCard key={product.id} product={product} />;
+            })}
+            <S.ObserverContainer ref={target} />
+          </S.ProductListContainer>
         </S.ProductContentWrapper>
       </CartContext.Provider>
     </>
